refactor: some formattings /  rename MantineNotifications -> messages
