+++ src/components/User/Comment/index.tsx
@@ -23,6 +23,7 @@ import { IconEditOff } from "@tabler/icons-react";
 import { remark } from "remark";
 import remarkBreaks from "remark-breaks";
 import remarkHtml from "remark-html";
+import { commentDeletedSuccessfulMsg } from "~/messages";
 
 import { useEffect, useState } from "react";
 import { toast } from "react-hot-toast";
@@ -32,7 +33,6 @@ import { useRouter } from "next/router";
 
 import LikeHeart from "~/components/Atom/LikeHeart";
 import UserAvatar from "~/components/Atom/UserAvatar";
-import { commentDeletedSuccessfulMsg } from "~/components/MantineNotifications";
 
 import { type Comment, Locale } from "~/types";
