+++ src/pages/grow/[reportId]/update/[postId]/index.tsx
@@ -9,6 +9,7 @@ import { useMediaQuery } from "@mantine/hooks";
 import { notifications } from "@mantine/notifications";
 import { IconChevronLeft } from "@tabler/icons-react";
 import dayjs from "dayjs";
+import { noPostAtThisDay } from "~/messages";
 
 import { useEffect, useState } from "react";
 
@@ -22,7 +23,6 @@ import Head from "next/head";
 import Link from "next/link";
 import { useRouter } from "next/router";
 
-import { noPostAtThisDay } from "~/components/MantineNotifications";
 import { PostCard } from "~/components/Post/Card";
 import PostsDatePicker from "~/components/Post/Datepicker";
